chore: configured logger and errors

+++ Core/Inc/lem.h
@@ -5,7 +5,7 @@
 #include <inttypes.h>
 
 void lem_init(void);
-uint16_t lem_get_current(void);
+float lem_get_current(void);
 void lem_new_value(uint16_t val);
 void lem_adc_callback(ADC_HandleTypeDef *hadc);
